search and table updates
