+++ README.md
@@ -28,7 +28,7 @@ Faculty: Prof. Mike Wu
     - Tabulated windows
       - Listings
       - Events/Landmarks/Restaurants
-6.  Admin Screen (K)
+6.  Admin Screen (A)
     - Tabulated windows for listings/events (CRUD)
 7.  Account (K)
     - Profile details (CRUD)
