--- client/src/api/locations.ts
@@ -1,4 +1,0 @@
-import api from "./api";
-
-export const getTrendingLocations = (config = {}) =>
-  api.get("trending-locations", config).then((res) => res.data);
