+++ client/src/components/index.ts
@@ -2,3 +2,4 @@ export { default as DisplayCard } from "./DisplayCard";
 export { default as LoadingSpinner } from "./LoadingSpinner";
 export { default as Navbar } from "./Navbar";
 export { default as Search } from "./Search";
+export { default as Table } from "./Table";
