+++ client/src/lib/types.ts
@@ -1,3 +1,5 @@
+import { Moment } from "moment";
+
 export type TLocation = {
   label: string;
   navigateTo: string;
@@ -28,4 +30,13 @@ export type TSignup = {
   fullname: string;
   email: string;
   password: string;
+};
+
+export type TSearch = {
+  location?: string;
+  numGuests?: number;
+  checkInDate?: string | Moment | null;
+  checkOutDate?: string | Moment | null;
+  budget?: string;
+  filters?: string[];
 };
